added precision

+++ ivy/functional/backends/paddle/experimental/statistical.py
@@ -376,11 +376,13 @@ def igamma(
 ) -> paddle.Tensor:
     results = []
     for ai, xi in zip(a.flatten(), x.flatten()):
+        ai = ai.astype("float64")
+        xi = xi.astype("float64")
 
         def integrand(t):
             return paddle.exp(-t) * paddle.pow(t, ai - 1)
 
-        intervals = paddle.linspace(0, xi, 10001)
+        intervals = paddle.linspace(0, xi, 10001).astype("float64")
         interval_width = xi / 10000
         values = integrand(intervals)
         integral = paddle.multiply((values[:-1] + values[1:]) / 2, interval_width)
